Add files via upload
Adding All America Micromouse Competition (AAMC Files) from the IEEE at UCLA maze bank. File format is different than those of the original repository as the format followed on this (https://github.com/mackorone/mms) maze solving simulator used in our program, but can be updated as needed to fit the rest of the repository. All of these mazes are classic sized, with start boxes at the bottom left corner, and the middle 4 boxes as the goal.